// src/app/modules/package/models/package.ts
export interface Package {
  id: string;
  ID: string;
  groupLeadName: string;
  groupLeadPhone: number;
  itineraryID: string;
  localAirportID: string;
  mutawefName: string;
  mutawefPhone: number;
  mutawefPicture: null;
  packageAvailableSeats: number;
  packageCapacity: string;
  packageDepartureDate: string;
  packageReturnDate: string;
  packageSeasonID: string;
  roomPriceAdult: number;
  roomPriceKids: number;
  roomQuantity: number;
  roomTypeID: string;
  roomTypeInfants: number;
}